+++ src/hooks/useHover.js
@@ -1,4 +1,4 @@
-import { useRef, useState, useEffect } from 'react';
+import { useRef, useState, useEffect } from 'react'
 
 export function useHover() {
   const [value, setValue] = useState(false)
@@ -8,21 +8,18 @@ export function useHover() {
   const handleMouseOver = () => setValue(true)
   const handleMouseOut = () => setValue(false)
 
-  useEffect(
-    () => {
-      const node = ref.current
-      if (node) {
-        node.addEventListener('mouseover', handleMouseOver)
-        node.addEventListener('mouseout', handleMouseOut)
+  useEffect(() => {
+    const node = ref.current
+    if (node) {
+      node.addEventListener('mouseover', handleMouseOver)
+      node.addEventListener('mouseout', handleMouseOut)
 
-        return () => {
-          node.removeEventListener('mouseover', handleMouseOver)
-          node.removeEventListener('mouseout', handleMouseOut)
-        }
+      return () => {
+        node.removeEventListener('mouseover', handleMouseOver)
+        node.removeEventListener('mouseout', handleMouseOut)
       }
-    },
-    [ref.current] // Recall only if ref changes
-  )
+    }
+  })
 
   return [ref, value]
 }
